Move auth status to AuthService
